ml-service/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
numpy==2.0.2
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
psycopg2-binary==2.9.9
